Server/requirements.txt
Flask
pymysql
cryptography
pytest
coverage
Flask-HTTPAuth
PyJWT
mysqlclient
flask-cors